Cilk: Use cleanup to create and release helper stack frames
While emitting a cilk spawn, there is a big difference between
the parent and helper function on creating stack frames.

For the parent function, on allocating the stack frame, we initialize
the stack frame immediately, and push the cleanup to release this stack
frame.

For the helper function, on allocating the stack frame, we cannot
initialize the stack frame, which should happen after the evaluation of
arguments and the evaluation of the left hand side. For example,

          f0() = _Cilk_spawn f2( f1() );

The evaluation sequence would be:

(0) Allocate the stack frame and initialize its worker to null
(1) Evaluate the left hand side, f0()
(2) Evaluate the argument, f1()
(3) Initialize the stack frame and detach
(4) Call the spawned function, f2()
(5) Do the assignment
(6) Destruct objects in reversed order from (5), (2), (1)
(7) Release the stack frame for the helper

Note that (7) will check if the worker is null or not. If it is not
null then do the cleanup, otherwise it is a noop.

Note that if the runtime __cilkrts_pop_frame and __cilkrts_leave_frame
can detect if a stack frame is uninitialized and act properly,
then __cilk_reset_worker call is not needed.