Merge captured statements with spawn
This patch changes spawn from being a captured statement
(DeprecatedCapturedStmt in this patch) to containing one. It also
removes the old ParallelRegionScopeInfo, and IsParallelRegion code.

There are now a couple of virtual methods in the CGCapturedStmtInfo that
allow customizing the codegen of captured statements.  For example, the
name of the helper function can be changed, and a prologue may be
emitted into the helper function.