Spawn exception-handling fixes
When the Cilk runtime begins or ends a spawned closure (do_work in the
runtime), it asserts that std::uncaught_exception() is false.

This patch makes three changes to exception handling that address cases
where this assertion might trigger:

1) Move the sync at the end of a try-statement from being a cleanup on
the try body to a cleanup around both the try and catches.  If we sync
before the catches, then std::uncaught_exception will be true when we
throw inside the try (but not in a spawned function). Consistent with
GCC, we now produce something like this:

try {
  try-block;
}  catch-blocks...
} finally {
  _Cilk_sync;
}

2) Move the sync before a throw-expression before the call to
__cxa_allocate_exception, since after the allocation returns,
std::uncaught_exception() will be true.

3) We were placing the excepting sync before __cxa_begin_catch, but
at that point std::uncaught_exception() is true.