Use clang_compile to build the cilk runtime
Setting CMAKE_C_COMPILER in the CMake file did not work with Ninja.

Also, create a variable CILKRTS_DEBUG to determine whether to set _DEBUG
or NDEBUG in the cilk runtime. We don't follow the setting from
CMAKE_BUILD_TYPE, because that will cause the runtime to emit debug
output if assertions are enabled in the clang build.