Fix exception handling issues with spawn
This new ABI is based on reading assembly output from ICC, and appears to
fix our known issues with exception handling.  These changes should be
incorporated into the next revision of the ABI specification.

In order to support calling pop and leave frame after a sync that
rethrows an exception, __cilk_sync is no longer marked as nounwind, and
is invoked rather than called as appropriate.  This required a large
number of test changes.

Also, the 'excepting sync' is not only called before catches, but is
called any time there is an exception in a spawn.